requirements.txt
google-generativeai
pytest
Flask
PyYAML
